refactor(api): moved PATCH and DELETE endpoints for orders to parametric endpoint
